backend/focus-sense-service/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.8.0
python-multipart==0.0.9
